requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
PyYAML
junos-eznc
PyMySQL